Alti2Reader: Properly display Post-Y2K jump dates
Internally, the N3 seems to store the month/year of each
jump as the number of months since Jan 2007. Since this
is represented as a 7-bit number, the "Y2K update" bumped
this offset to Jan 2015. Thus, update the date conversion
code to properly display jump dates from an N3 that has
had the Y2K update installed.

Signed-off-by: evilwombat <evilwombat@server.fake>