Alti2Reader: Don't choke on invalid dates
If we read data stored on a Pre-Y2K unit that's been jumped
after the rollover date, it is possible for the rolled-over
dates to be invalid due to the day exceeding the number of
days in the (now-incorrect) month. This would normally lead
to a crash, completely preventing the salvage of the data
being read. Instead, try to process and display the invalid
dates anyway, marking the invalid ones with "??", so that
at least other jump details may be salvaged.

In the future, we might introduce a "minimum valid date"
setting, which would allow any rolled-over jumps to be
corrected automatically.

Signed-off-by: evilwombat <evilwombat@server.fake>